+++ common/StringVector.h
@@ -2,6 +2,7 @@
 #define __STRINGVECTOR_H__
 
 #include <string>
+#include <cinttypes>
 
 class CStringVektor
 {
@@ -13,7 +14,7 @@ public:
 	CStringVektor();
 	~CStringVektor();
 	void Clean();
-	int AddInt(int nVal);
+	int AddInt(intptr_t nVal);
 	int AddString(char *szString,int nSize=-1);
 	int InsertString(int nIndex, char *szString,int nSize=-1);
 	int DeleteString(int nIndex);
@@ -55,7 +56,7 @@ public:
 	int InsertRow(int nRow);
 	int AddString(int nRow, char *szString, int nString=-1);
 	int AddString(int nRow, const std::string &szString, int nString=-1);
-	int AddInt(int nRow, int nVal);
+	int AddInt(int nRow, intptr_t nVal);
 	int DeleteRow(int nRow);
 	int DeleteColumn(int nColumn);
 	void Reset();
